Initial commit with hello world
Why learn ATS?

I've taken the long way, moving with the biased-random gait of a
bacterium. Slowly I've drifted between languages, never liking any of
them. Here and there I have seen glimmers of something beautiful. But my
code always topples. I build, but soon the rate of regression becomes
too high, the program becomes too complex to keep in my head. Of course,
I can test my code. But testing seems an unholy necromancy to me, where
the corpse of a program is proped up with steel joints and electricity.
A testing suite plays the role of the environment in biology. Where
deviations from what is tested cause the program to fail. The test suite
is a vessel wrapping the code, an adhoc specification. Test coverage is
never complete. This terrorfies me. The clean test result offers only
false assurance. I would rather have compile-time assurances that the
program is correct.

Eventually I found Haskell. And Haskell is a thing of beauty. For the
first time, I found I could write good code. Programs that type check
generally work. But still, something was missing. When I studied the
monad laws, I felt a creeping sense of unease. The definition of a monad
is only partially checked by the compiler. The `return` and `bind`
functions are typechecked, but still a false monad can be created where
the monadic laws are violated. The monadic laws are not specified in the
type signatures, but rather only in the ad hoc human realm. The same
goes for semigroups, which are types defined by the existence of a
closed, associative operator. The closed aspect of the semigroup is
cleanly expressed with the typeclass:

```
class Semigroup a where
    op :: a -> a -> a
```

But the associative aspect is left to the user. There is nothing to
prevent an irresponsible library writer from creating a false Semigroup
instance. I want to be able to write this:

```
class Semigroup a where
    op :: a -> a -> a | op (op x y) z == op x (op y z)
```

That is, I want to be able to expres the functions and all laws in code.
Haskell has a test suite, called Quickcheck, that can test properties
by passing random values and asserting the properties hold. But I want a
stronger guarantee.

I don't have a strong background in type theory. But from a little
reading, I think what I am looking for is a language with dependent
types. ATS seems to be the most well-developed such language (apart from
Coq, which seems to have no value as a practical programming language).

I tried to install ATS2 using the Archlinux rep (through yaourt) but
that failed for reasons I have no patience to explore. So I installed
from source, which went well. The hello world program on the ATS2
homepage compiled and ran without a hitch. There are a few oddities
about the syntax that I will need to explore, but my ATS2 system is up
and running. So we are in business.